+++ KretaParancsosiAlkalmazas/Repositories/SchoolClassesRepo.cs
@@ -22,7 +22,7 @@ namespace Kreta.Repositories
             }
         }
 
-        public int GetSchoolClassId(int grade, char gradeType)
+        public long GetSchoolClassId(int grade, char gradeType)
         {
             SchoolClass schoolClass = schoolClasses.Find(schoolClass => schoolClass.Grade == grade && schoolClass.GradeType == gradeType);
             return schoolClass.Id;
